add charsets function
